Removed stegos_crypto::hashcash
